Several changes bundled into this one. metric_t was changed from double to float to save on memory. Metadata generation was added (topography is a placeholder pending a full rework on topo metrics). A bug was fixed which caused DEM generation to use unreasonable amounts of memory if the input DEM files were large.

+++ src/run/ProductHandler.cpp
@@ -1,7 +1,6 @@
 #include"run_pch.hpp"
 #include"ProductHandler.hpp"
 #include"PointMetricCalculator.hpp"
-#include"..\logger\LapisLogger.hpp"
 
 namespace lapis {
 
@@ -23,7 +22,7 @@ namespace lapis {
 	void ProductHandler::deleteTempDirIfEmpty() const
 	{
 		namespace fs = std::filesystem;
-		if (fs::is_empty(tempDir())) {
+		if (fs::exists(tempDir()) && fs::is_empty(tempDir())) {
 			fs::remove(tempDir());
 		}
 	}
